Move har handlers out of main file
