+++ src/handlers/index.ts
@@ -1,7 +1,9 @@
 import * as auth from './auth'
 import * as cloud from './cloud'
+import * as har from './har'
 
 export function initialize() {
   auth.initialize()
   cloud.initialize()
+  har.initialize()
 }
